VFS, new file handling code and allocator improvements.
I wrote a primitive VFS to facilitate half-life style mods in directories.
Engine knows multiple mod directories that are subdirectories of ./data/
and can be added using add_mod_directory.

When opening a file, the VFS first looks for it in the most recently
added directory, then the second most recent, etc. There is always
one "main" directory (./data/main) that is required for the engine
to run and contains "default" data shared by all mods.

File names can alternatively use a specific notation (::) to specify
exact mod directory to look in, e.g. mod1::file.txt will always look
for ./data/mod1/file.txt .

The VFS comes with a simple File struct to read, write and append to
files. Files are always read whole, when opened for reading, and when
writing/appending, changes are also written out in single call when
the file is closed. This should simplify and possible future
implementation of an archive format.

File handling also required a change in allocator code: a realloc()
function. Just as it sounds, realloc() reallocates arrays allocated
with alloc().