Improved Vector safety.
We now assert that Vector is not reallocated during foreach,
and the ptr() to element method has been removed.
Users can still get pointers to elements by e.g. &vector[elemIdx],
but a warning has been added in the documentation that this is not
safe and the pointer will be invalid after reallocation.